etl/requirements.txt
redis==5.0.7
requests==2.32.3
psycopg2-binary==2.9.6
backoff==2.2.1
elastic-transport==8.13.1
elasticsearch==8.14.0
python-dotenv==1.0.1

